refactor: guarantee the order in which rules are created and abstract behaviors
