+++ lib/utils/generateManifest/generateManifest.utils.js
@@ -3,7 +3,9 @@ import Ajv from 'ajv';
 import ajvErrors from 'ajv-errors';
 import addKeywords from 'ajv-keywords';
 
-import azionConfigSchema from './fixtures/schema.js';
+import azionConfigSchema from './helpers/schema.js';
+import { requestBehaviors, responseBehaviors } from './helpers/bahaviors.js';
+
 /**
  * Validates the provided configuration against a JSON Schema.
  * This function uses AJV (Another JSON Schema Validator) to validate the configuration.
@@ -35,8 +37,8 @@ function jsToJson(config) {
 
   const payloadCDN = {
     origin: [],
-    rules: [],
     cache: [],
+    rules: [],
   };
 
   // Helper function to safely evaluate mathematical expressions
@@ -101,8 +103,28 @@ function jsToJson(config) {
     });
   }
 
-  // Convert request rules
+  // Helper function to add behaviors to a rule
+  const addBehaviors = (cdnRule, behaviors, behaviorDefinitions) => {
+    if (behaviors && typeof behaviors === 'object') {
+      Object.entries(behaviors).forEach(([key, value]) => {
+        if (behaviorDefinitions[key]) {
+          const transformedBehavior = behaviorDefinitions[key].transform(
+            value,
+            payloadCDN,
+          );
+          if (Array.isArray(transformedBehavior)) {
+            cdnRule.behaviors.push(...transformedBehavior);
+          } else if (transformedBehavior) {
+            cdnRule.behaviors.push(transformedBehavior);
+          }
+        } else {
+          console.warn(`Unknown behavior: ${key}`);
+        }
+      });
+    }
+  };
 
+  // Convert request rules
   config?.rules?.request?.forEach((rule, index) => {
     const cdnRule = {
       name: rule.name,
@@ -113,7 +135,6 @@ function jsToJson(config) {
       criteria: [
         [
           {
-            // eslint-disable-next-line no-template-curly-in-string
             variable: `\${${rule.variable ?? 'uri'}}`,
             operator: 'matches',
             conditional: 'if',
@@ -121,194 +142,15 @@ function jsToJson(config) {
           },
         ],
       ],
-      behaviors: rule.behaviors ? [...rule.behaviors] : [],
+      behaviors: [],
     };
 
-    if (rule.description) {
-      cdnRule.description = rule.description;
-    }
-
-    // Add specific behaviors for the rewrite rule
-    if (rule.rewrite) {
-      const functionString = rule.rewrite.set.toString();
-      const paramNameMatch = functionString.match(/\(([^)]+)\)/);
-      let paramName = paramNameMatch ? paramNameMatch[1] : null;
-      if (paramName) {
-        // eslint-disable-next-line prefer-destructuring
-        paramName = paramName.trim().split(/\s*,\s*/)[0];
-      }
-
-      if (rule.rewrite.match) {
-        cdnRule.behaviors.push({
-          name: 'capture_match_groups',
-          target: {
-            captured_array: paramName || 'captured',
-            // eslint-disable-next-line no-template-curly-in-string
-            variable: `\${${rule.variable ?? 'uri'}}`,
-            regex: rule.rewrite.match,
-          },
-        });
-      }
-      // Transform the set function into a replacement string
-      const pathTransformation = functionString
-        .replace(/.*=>\s*`/, '') // Remove the function part up to the template string
-        .replace(/`\s*;?\s*}?\s*$/, '') // Removes the end of the template string and possible function closures
-        .replace(/\$\{([^}]+)\}/g, (match, p1) => `\${${p1}}`); // Replace ${other[index]} with %{other[index]}
-
-      cdnRule.behaviors.push({
-        name: 'rewrite_request',
-        target: pathTransformation,
-      });
-    }
-
-    // Cache handling within rules
-    if (typeof rule.cache === 'string') {
-      cdnRule.behaviors.push({
-        name: 'set_cache_policy',
-        target: rule.cache,
-      });
-    } else if (typeof rule.cache === 'object') {
-      // Adds the cache object to the cache array
-      const cacheSetting = {
-        name: rule.cache?.name,
-        ...rule.cache, // Here you can expand or adjust fields as necessary
-      };
-      payloadCDN.cache.push(cacheSetting);
-      // Adds the set_cache_policy behavior with the generated cache name
-      cdnRule.behaviors.push({
-        name: 'set_cache_policy',
-        target: rule.cache?.name,
-      });
-      // Updates the rule to use the generated cache name
-      // eslint-disable-next-line no-param-reassign
-      rule.cache = rule.cache?.name;
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rule, 'forwardCookies')) {
-      if (rule.forwardCookies) {
-        cdnRule.behaviors.push({
-          name: 'forward_cookies',
-          target: null,
-        });
-      }
-    }
-
-    if (rule.setOrigin) {
-      const origin = payloadCDN.origin.find(
-        (o) =>
-          o.name === rule.setOrigin.name &&
-          o.origin_type === rule.setOrigin.type,
-      );
-
-      if (!origin) {
-        throw new Error(
-          `Rule setOrigin name '${rule.setOrigin.name}' not found in the origin settings`,
-        );
-      } else if (origin.origin_type !== rule.setOrigin.type) {
-        throw new Error(
-          `Rule setOrigin originType '${rule.setOrigin.type}' does not match the origin settings`,
-        );
-      }
-
-      const originBehavior = {
-        name: 'set_origin',
-        target: origin.name,
-      };
-
-      cdnRule.behaviors.push(originBehavior);
-    }
-
-    if (rule.runFunction) {
-      const runFunctionBehavior = {
-        name: 'run_function',
-        target: rule.runFunction.path,
-      };
-
-      cdnRule.behaviors.push(runFunctionBehavior);
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rule, 'enableGZIP')) {
-      if (rule.enableGZIP) {
-        cdnRule.behaviors.push({
-          name: 'enable_gzip',
-          target: '',
-        });
-      }
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rule, 'bypassCache')) {
-      if (rule.bypassCache) {
-        cdnRule.behaviors.push({
-          name: 'bypass_cache_phase',
-          target: null,
-        });
-      }
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rule, 'httpToHttps')) {
-      if (rule.httpToHttps) {
-        cdnRule.behaviors.push({
-          name: 'redirect_http_to_https',
-          target: null,
-        });
-      }
-    }
-
-    if (rule.redirectTo301) {
-      cdnRule.behaviors.push({
-        name: 'redirect_to_301',
-        target: rule.redirectTo301,
-      });
-    }
-
-    if (rule.redirectTo302) {
-      cdnRule.behaviors.push({
-        name: 'redirect_to_302',
-        target: rule.redirectTo302,
-      });
-    }
-
-    if (rule.setCookie) {
-      const cookieBehavior = {
-        name: 'add_request_cookie',
-        target: rule.setCookie,
-      };
-      cdnRule.behaviors.push(cookieBehavior);
-    }
-
-    if (rule.setHeaders) {
-      rule.setHeaders.forEach((header) => {
-        cdnRule.behaviors.push({
-          name: 'add_request_header',
-          target: header,
-        });
-      });
-    }
-
-    if (rule.capture) {
-      cdnRule.behaviors.push({
-        name: 'capture_match_groups',
-        target: {
-          regex: rule.capture.match,
-          captured_array: rule.capture.captured,
-          subject: `\${${rule.capture.subject ?? 'uri'}}`,
-        },
-      });
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rule, 'deliver')) {
-      if (rule.deliver) {
-        cdnRule.behaviors.push({
-          name: 'deliver',
-          target: null,
-        });
-      }
-    }
+    addBehaviors(cdnRule, rule.behavior, requestBehaviors);
 
     payloadCDN.rules.push(cdnRule);
   });
 
-  // response rules
+  // Convert response rules
   if (config?.rules?.response) {
     config.rules.response.forEach((rule, index) => {
       const cdnRule = {
@@ -327,88 +169,10 @@ function jsToJson(config) {
             },
           ],
         ],
-        behaviors: rule.behaviors ? [...rule.behaviors] : [],
+        behaviors: [],
       };
 
-      if (rule.setCookie) {
-        cdnRule.behaviors.push({
-          name: 'add_response_cookie',
-          target: rule.setCookie,
-        });
-      }
-
-      if (rule.setHeaders) {
-        rule.setHeaders.forEach((header) => {
-          cdnRule.behaviors.push({
-            name: 'add_response_header',
-            target: header,
-          });
-        });
-      }
-
-      if (Object.prototype.hasOwnProperty.call(rule, 'enableGZIP')) {
-        if (rule.enableGZIP) {
-          cdnRule.behaviors.push({
-            name: 'enable_gzip',
-            target: '',
-          });
-        }
-      }
-
-      if (rule.filterCookie) {
-        cdnRule.behaviors.push({
-          name: 'filter_response_cookie',
-          target: rule.filterCookie,
-        });
-      }
-
-      if (rule.filterHeader) {
-        cdnRule.behaviors.push({
-          name: 'filter_response_header',
-          target: rule.filterHeader,
-        });
-      }
-
-      if (rule.runFunction) {
-        cdnRule.behaviors.push({
-          name: 'run_function',
-          target: rule.runFunction.path,
-        });
-      }
-
-      if (rule.redirectTo301) {
-        cdnRule.behaviors.push({
-          name: 'redirect_to_301',
-          target: rule.redirectTo301,
-        });
-      }
-
-      if (rule.redirectTo302) {
-        cdnRule.behaviors.push({
-          name: 'redirect_to_302',
-          target: rule.redirectTo302,
-        });
-      }
-
-      if (rule.capture) {
-        cdnRule.behaviors.push({
-          name: 'capture_match_groups',
-          target: {
-            regex: rule.capture.match,
-            captured_array: rule.capture.captured,
-            subject: `\${${rule.capture.subject ?? 'uri'}}`,
-          },
-        });
-      }
-
-      if (Object.prototype.hasOwnProperty.call(rule, 'deliver')) {
-        if (rule.deliver) {
-          cdnRule.behaviors.push({
-            name: 'deliver',
-            target: null,
-          });
-        }
-      }
+      addBehaviors(cdnRule, rule.behavior, responseBehaviors);
 
       payloadCDN.rules.push(cdnRule);
     });
