+++ lib/utils/generateManifest/generateManifest.utils.test.js
@@ -29,10 +29,12 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/test',
-              cache: {
-                name: 'directCache',
-                browser_cache_settings_maximum_ttl: 300,
-                cdn_cache_settings_maximum_ttl: 600,
+              behavior: {
+                cache: {
+                  name: 'directCache',
+                  browser_cache_settings_maximum_ttl: 300,
+                  cdn_cache_settings_maximum_ttl: 600,
+                },
               },
             },
           ],
@@ -55,14 +57,16 @@ describe('Utils - generateManifest', () => {
             {
               name: 'rewriteRule',
               match: '/path',
-              capture: {
-                match: '^(./)([^/])$',
-                captured: 'other',
-                subject: 'uri',
-              },
-              rewrite: {
-                match: '^(./)([^/])$',
-                set: (other) => `/${other[0]}/${other[1]}`,
+              behavior: {
+                capture: {
+                  match: '^(./)([^/])$',
+                  captured: 'other',
+                  subject: 'uri',
+                },
+                rewrite: {
+                  match: '^(./)([^/])$',
+                  set: (other) => `/${other[0]}/${other[1]}`,
+                },
               },
             },
           ],
@@ -120,8 +124,10 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/no-cache',
-              runFunction: {
-                path: '.edge/worker.js',
+              behavior: {
+                runFunction: {
+                  path: '.edge/worker.js',
+                },
               },
             },
           ],
@@ -147,9 +153,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'object_storage',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'object_storage',
+                },
               },
             },
           ],
@@ -224,7 +232,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/',
-              forwardCookies: true,
+              behavior: {
+                forwardCookies: true,
+              },
             },
           ],
         },
@@ -249,7 +259,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/no-forward',
-              forwardCookies: false,
+              behavior: {
+                forwardCookies: false,
+              },
             },
             {
               name: 'testRule',
@@ -277,6 +289,7 @@ describe('Utils - generateManifest', () => {
         ]),
       );
     });
+
     it('should correctly process the setOrigin rule with all optional fields provided', () => {
       const azionConfigWithAllFields = {
         origin: [
@@ -292,9 +305,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/_next',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'object_storage',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'object_storage',
+                },
               },
             },
           ],
@@ -318,9 +333,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/_next_no_type',
-              setOrigin: {
-                name: 'nameWithoutType',
-                // type is missing
+              behavior: {
+                setOrigin: {
+                  name: 'nameWithoutType',
+                  // type is missing
+                },
               },
             },
           ],
@@ -354,8 +371,10 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/run-function-test-path-only',
-              runFunction: {
-                path: '.edge/worker.js',
+              behavior: {
+                runFunction: {
+                  path: '.edge/worker.js',
+                },
               },
             },
           ],
@@ -373,20 +392,23 @@ describe('Utils - generateManifest', () => {
         ]),
       );
     });
+
     it('should include the behavior deliver when deliver is true', () => {
-      const azionConfigComDeliver = {
+      const azionConfig = {
         rules: {
           request: [
             {
               name: 'testRule',
               match: '/path',
-              deliver: true,
+              behavior: {
+                deliver: true,
+              },
             },
           ],
         },
       };
 
-      const result = jsToJson(azionConfigComDeliver);
+      const result = jsToJson(azionConfig);
       expect(result.rules[0].behaviors).toEqual(
         expect.arrayContaining([
           expect.objectContaining({
@@ -395,6 +417,7 @@ describe('Utils - generateManifest', () => {
         ]),
       );
     });
+
     it('should throw an error if deliver is not a boolean', () => {
       const azionConfig = {
         rules: {
@@ -402,7 +425,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/path',
-              deliver: 'true', // Incorretamente definido como string
+              behavior: {
+                deliver: 'true', // Incorretamente definido como string
+              },
             },
           ],
         },
@@ -420,7 +445,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/',
-              setCookie: true, // Incorretamente definido como boolean
+              behavior: {
+                setCookie: true, // Incorretamente definido como boolean
+              },
             },
           ],
         },
@@ -438,7 +465,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/',
-              setHeaders: { key: 'value' }, // Incorretamente definido como objeto
+              behavior: {
+                setHeaders: { key: 'value' }, // Incorretamente definido como objeto
+              },
             },
           ],
         },
@@ -456,7 +485,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/path',
-              deliver: true,
+              behavior: {
+                deliver: true,
+              },
             },
           ],
         },
@@ -479,7 +510,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/',
-              setCookie: 'sessionId=abc123',
+              behavior: {
+                setCookie: 'sessionId=abc123',
+              },
             },
           ],
         },
@@ -503,7 +536,9 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/',
-              setHeaders: ['Authorization: Bearer abc123'], // Corretamente definido como uma array de strings
+              behavior: {
+                setHeaders: ['Authorization: Bearer abc123'], // Corretamente definido como uma array de strings
+              },
             },
           ],
         },
@@ -537,9 +572,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'object_storage',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'object_storage',
+                },
               },
             },
           ],
@@ -575,9 +612,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'another origin storage',
-                type: 'object_storage',
+              behavior: {
+                setOrigin: {
+                  name: 'another origin storage',
+                  type: 'object_storage',
+                },
               },
             },
           ],
@@ -604,9 +643,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'another_type',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'another_type',
+                },
               },
             },
           ],
@@ -625,9 +666,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'object_storage',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'object_storage',
+                },
               },
             },
           ],
@@ -655,9 +698,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my origin storage',
-                type: 'another_type',
+              behavior: {
+                setOrigin: {
+                  name: 'my origin storage',
+                  type: 'another_type',
+                },
               },
             },
           ],
@@ -684,9 +729,11 @@ describe('Utils - generateManifest', () => {
             {
               name: 'testRule',
               match: '/api',
-              setOrigin: {
-                name: 'my single origin',
-                type: 'single_origin',
+              behavior: {
+                setOrigin: {
+                  name: 'my single origin',
+                  type: 'single_origin',
+                },
               },
             },
           ],
@@ -718,6 +765,7 @@ describe('Utils - generateManifest', () => {
       );
     });
   });
+
   it('should correctly handle bypassCache behavior', () => {
     const azionConfig = {
       rules: {
@@ -725,7 +773,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testBypassCache',
             match: '/',
-            bypassCache: true,
+            behavior: {
+              bypassCache: true,
+            },
           },
         ],
       },
@@ -741,6 +791,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly handle redirect to 301', () => {
     const azionConfig = {
       rules: {
@@ -748,7 +799,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testRedirect301',
             match: '/',
-            redirectTo301: 'https://example.com',
+            behavior: {
+              redirectTo301: 'https://example.com',
+            },
           },
         ],
       },
@@ -772,7 +825,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testRedirect302',
             match: '/',
-            redirectTo302: 'https://example.com',
+            behavior: {
+              redirectTo302: 'https://example.com',
+            },
           },
         ],
       },
@@ -788,6 +843,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly handle capture match groups', () => {
     const azionConfig = {
       rules: {
@@ -795,10 +851,12 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testCapture',
             match: '/',
-            capture: {
-              match: '^/user/(.*)',
-              captured: 'userId',
-              subject: 'uri',
+            behavior: {
+              capture: {
+                match: '^/user/(.*)',
+                captured: 'userId',
+                subject: 'uri',
+              },
             },
           },
         ],
@@ -820,6 +878,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly handle filterCookie behavior', () => {
     const azionConfig = {
       rules: {
@@ -827,7 +886,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testFilterCookie',
             match: '/',
-            filterCookie: '_cookie',
+            behavior: {
+              filterCookie: '_cookie',
+            },
           },
         ],
       },
@@ -843,6 +904,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly process rules in the response phase', () => {
     const azionConfig = {
       rules: {
@@ -850,7 +912,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testResponsePhase',
             match: '/',
-            setHeaders: ['X-Test-Header: value'],
+            behavior: {
+              setHeaders: ['X-Test-Header: value'],
+            },
           },
         ],
       },
@@ -866,6 +930,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly add multiple response headers', () => {
     const azionConfig = {
       rules: {
@@ -873,10 +938,12 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testMultipleHeaders',
             match: '/',
-            setHeaders: [
-              'X-Frame-Options: DENY',
-              "Content-Security-Policy: default-src 'self'",
-            ],
+            behavior: {
+              setHeaders: [
+                'X-Frame-Options: DENY',
+                "Content-Security-Policy: default-src 'self'",
+              ],
+            },
           },
         ],
       },
@@ -896,6 +963,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should correctly handle enableGZIP behavior', () => {
     const azionConfig = {
       rules: {
@@ -903,7 +971,9 @@ describe('Utils - generateManifest', () => {
           {
             name: 'testEnableGZIP',
             match: '/',
-            enableGZIP: true,
+            behavior: {
+              enableGZIP: true,
+            },
           },
         ],
       },
@@ -919,6 +989,7 @@ describe('Utils - generateManifest', () => {
       ]),
     );
   });
+
   it('should handle rules with description and active properties correctly', () => {
     const azionConfig = {
       rules: {
@@ -932,6 +1003,7 @@ describe('Utils - generateManifest', () => {
           {
             name: 'Second Rule',
             match: '/api',
+            behavior: {},
             // description is not provided here
             active: true,
           },
@@ -939,6 +1011,7 @@ describe('Utils - generateManifest', () => {
             name: 'Third Rule',
             match: '/home',
             description: 'This rule handles home traffic.',
+            behavior: {},
             // active is not provided here
           },
         ],
@@ -969,12 +1042,12 @@ describe('Utils - generateManifest', () => {
     const azionConfig = {
       rules: {
         request: [
-          { name: 'First Request Rule', match: '/' },
-          { name: 'Second Request Rule', match: '/second' },
+          { name: 'First Request Rule', match: '/', behavior: {} },
+          { name: 'Second Request Rule', match: '/second', behavior: {} },
         ],
         response: [
-          { name: 'First Response Rule', match: '/' },
-          { name: 'Second Response Rule', match: '/second' },
+          { name: 'First Response Rule', match: '/', behavior: {} },
+          { name: 'Second Response Rule', match: '/second', behavior: {} },
         ],
       },
     };
@@ -1003,6 +1076,72 @@ describe('Utils - generateManifest', () => {
         name: 'Second Response Rule',
         order: 3,
       }),
+    );
+  });
+
+  it('should maintain the order of behaviors as specified by the user', () => {
+    const azionConfig = {
+      origin: [
+        {
+          name: 'my origin storage',
+          type: 'object_storage',
+          bucket: 'mybucket',
+          prefix: 'myfolder',
+        },
+      ],
+      rules: {
+        request: [
+          {
+            name: 'testRule',
+            match: '/',
+            behavior: {
+              setHeaders: ['Authorization: Bearer abc123'],
+              deliver: true,
+              setOrigin: {
+                name: 'my origin storage',
+                type: 'object_storage',
+              },
+            },
+          },
+        ],
+      },
+    };
+
+    const result = jsToJson(azionConfig);
+    expect(result.rules[0].behaviors).toEqual([
+      expect.objectContaining({
+        name: 'add_request_header',
+        target: 'Authorization: Bearer abc123',
+      }),
+      expect.objectContaining({
+        name: 'deliver',
+      }),
+      expect.objectContaining({
+        name: 'set_origin',
+        target: 'my origin storage',
+      }),
+    ]);
+  });
+  it('should throw an error when the origin settings are not defined', () => {
+    const azionConfigWithoutOrigin = {
+      rules: {
+        request: [
+          {
+            name: 'testRule',
+            match: '/api',
+            behavior: {
+              setOrigin: {
+                name: 'undefined origin',
+                type: 'object_storage',
+              },
+            },
+          },
+        ],
+      },
+    };
+
+    expect(() => jsToJson(azionConfigWithoutOrigin)).toThrow(
+      "Rule setOrigin name 'undefined origin' not found in the origin settings",
     );
   });
 });
