+++ lib/utils/generateManifest/helpers/azion.config.example.js
@@ -32,14 +32,16 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/rewrite$',
-        cache: 'mycache1',
-        rewrite: {
-          match: '^(./)([^/])$',
-          set: (captured) => `/new/${captured[1]}`, // Rewrites /original/image.jpg to /new/image.jpg
+        behavior: {
+          cache: 'mycache1',
+          rewrite: {
+            match: '^(./)([^/])$',
+            set: (captured) => `/new/${captured[1]}`, // Rewrites /original/image.jpg to /new/image.jpg
+          },
+          setCookie: 'user=12345; Path=/; Secure',
+          setHeaders: 'Cache-Control: no-cache',
+          forwardCookies: true,
         },
-        setCookie: 'user=12345; Path=/; Secure',
-        setHeaders: 'Cache-Control: no-cache',
-        forwardCookies: true,
       },
       {
         name: 'staticContentRuleExample',
@@ -48,11 +50,13 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/_statics/',
-        setOrigin: {
-          name: 'myneworigin',
-          type: 'object_storage',
+        behavior: {
+          setOrigin: {
+            name: 'myneworigin',
+            type: 'object_storage',
+          },
+          deliver: true,
         },
-        deliver: true,
       },
       {
         name: 'computeFunctionRuleExample',
@@ -60,8 +64,10 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/compute/',
-        runFunction: {
-          path: '.edge/worker.js',
+        behavior: {
+          runFunction: {
+            path: '.edge/worker.js',
+          },
         },
       },
       {
@@ -70,7 +76,9 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/old-url$',
-        redirectTo301: 'https://newsite.com/new-url',
+        behavior: {
+          redirectTo301: 'https://newsite.com/new-url',
+        },
       },
       {
         name: 'gzipCompressionRuleExample',
@@ -78,7 +86,9 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/compress',
-        enableGZIP: true,
+        behavior: {
+          enableGZIP: true,
+        },
       },
       {
         name: 'apiHeaderRuleExample',
@@ -86,7 +96,9 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/api',
-        setHeaders: ['X-API-Version: 1', 'X-Frame-Options: deny'],
+        behavior: {
+          setHeaders: ['X-API-Version: 1', 'X-Frame-Options: deny'],
+        },
       },
       {
         name: 'cookieSettingRuleExample',
@@ -94,7 +106,9 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/check',
-        setCookie: 'test=12345; Path=/; Secure; HttpOnly',
+        behavior: {
+          setCookie: 'test=12345; Path=/; Secure; HttpOnly',
+        },
       },
       {
         name: 'userCaptureRuleExample',
@@ -102,10 +116,12 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/user/(.*)',
-        capture: {
-          regex: '^(.*)$',
-          captured: 'user_id',
-          subject: 'uri',
+        behavior: {
+          capture: {
+            regex: '^(.*)$',
+            captured: 'user_id',
+            subject: 'uri',
+          },
         },
       },
       {
@@ -114,19 +130,21 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/some-path',
-        cache: {
-          name: 'cacheDinamico',
-          stale: true,
-          queryStringSort: true,
-          methods: {
-            post: true,
-            options: true,
-          },
-          browser: {
-            maxAgeSeconds: 3600, // 1 hour
-          },
-          edge: {
-            maxAgeSeconds: 600, // 10 minutes
+        behavior: {
+          cache: {
+            name: 'cacheDinamico',
+            stale: true,
+            queryStringSort: true,
+            methods: {
+              post: true,
+              options: true,
+            },
+            browser: {
+              maxAgeSeconds: 3600, // 1 hour
+            },
+            edge: {
+              maxAgeSeconds: 600, // 10 minutes
+            },
           },
         },
       },
@@ -137,7 +155,9 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/bypass',
-        bypassCache: true,
+        behavior: {
+          bypassCache: true,
+        },
       },
       {
         name: 'forceHttpsRuleExample',
@@ -145,7 +165,9 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/secure-area',
-        redirectHttps: true,
+        behavior: {
+          httpToHttps: true,
+        },
       },
       {
         name: 'UriRedirectExample',
@@ -153,13 +175,15 @@ export default {
         active: true,
         match: '.*', // Captures all URIs
         variable: 'uri', // Defines the variable to be captured
-        capture: {
-          match: '^(.*)$', // Captures the entire URI path
-          captured: 'uri_path', // Name of the variable where the captured path will be stored
-          subject: 'uri', // Indicates that the capture will be made on the 'uri' variable
+        behavior: {
+          capture: {
+            match: '^(.*)$', // Captures the entire URI path
+            captured: 'uri_path', // Name of the variable where the captured path will be stored
+            subject: 'uri', // Indicates that the capture will be made on the 'uri' variable
+          },
+          redirectTo302: `https://example.com/%{uri_path}`, // Uses the captured path as part of the new URL
+          filterCookie: 'original_uri_cookie', // Removes the original cookie to avoid conflicts or duplicate information
         },
-        redirectTo302: `https://example.com/%{uri_path}`, // Uses the captured path as part of the new URL
-        filterCookie: 'original_uri_cookie', // Removes the original cookie to avoid conflicts or duplicate information
       },
     ],
     response: [
@@ -170,11 +194,13 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/api/data',
-        setHeaders: 'Content-Type: application/json',
-        setCookie: 'session=abcdef; Path=/; HttpOnly',
-        filterHeader: 'Server',
-        filterCookie: 'tracking',
-        enableGZIP: true,
+        behavior: {
+          setHeaders: 'Content-Type: application/json',
+          setCookie: 'session=abcdef; Path=/; HttpOnly',
+          filterHeader: 'Server',
+          filterCookie: 'tracking',
+          enableGZIP: true,
+        },
       },
       {
         name: 'userProfileRedirectRuleExample',
@@ -183,7 +209,9 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/user/profile',
-        redirectTo301: 'https://newsite.com/profile',
+        behavior: {
+          redirectTo301: 'https://newsite.com/profile',
+        },
       },
       {
         name: 'computeResultFunctionRuleExample',
@@ -192,17 +220,19 @@ export default {
         active: true,
         variable: 'uri', // Optional, defaults to 'uri' if not provided
         match: '^/compute-result',
-        runFunction: {
-          path: '.edge/computeResult.js',
+        behavior: {
+          runFunction: {
+            path: '.edge/computeResult.js',
+          },
+          // This rule captures the full URI path and stores it in a variable named 'full_path_arr'.
+          capture: {
+            match: '^(.*)$', // The regular expression '^(.*)$' captures the entire URI path.
+            captured: 'full_path_arr', // The result of the capture is stored in the variable 'full_path_arr'.
+            subject: 'uri', // The capture is based on the value of the 'uri' variable.
+          },
+          // Permanently redirects to the first element captured in 'full_path_arr'.
+          redirectTo301: '%{full_path_arr[0]}', // Uses the first element of the 'full_path_arr' array as part of the new URL.
         },
-        // This rule captures the full URI path and stores it in a variable named 'full_path_arr'.
-        capture: {
-          match: '^(.*)$', // The regular expression '^(.*)$' captures the entire URI path.
-          captured: 'full_path_arr', // The result of the capture is stored in the variable 'full_path_arr'.
-          subject: 'uri', // The capture is based on the value of the 'uri' variable.
-        },
-        // Permanently redirects to the first element captured in 'full_path_arr'.
-        redirectTo301: '%{full_path_arr[0]}', // Uses the first element of the 'full_path_arr' array as part of the new URL.
       },
       {
         name: 'userProfileRedirectRuleExample',
@@ -211,7 +241,9 @@ export default {
         // eslint-disable-next-line no-template-curly-in-string
         variable: 'cookie_name', // Example using cookie value
         match: '^user-profile$', // Matches based on the cookie value
-        redirectTo301: 'https://newsite.com/profile',
+        behavior: {
+          redirectTo301: 'https://newsite.com/profile',
+        },
       },
       {
         name: 'temporaryPageRedirectRuleExample',
@@ -221,7 +253,9 @@ export default {
         // eslint-disable-next-line no-template-curly-in-string
         variable: 'args', // All query parameters
         match: '^old-page$', // Matches based on the presence of specific query parameters
-        redirectTo302: 'https://newsite.com/new-page',
+        behavior: {
+          redirectTo302: 'https://newsite.com/new-page',
+        },
       },
     ],
   },
